1. Updated the BSP for i.MX RT1050 2. Added the Keil Workspace 3. Changed the hello_world.c for RT1050

+++ platform/vendor_bsp/nxp/MIMXRT1052/drivers/fsl_ewm.h
@@ -1,35 +1,9 @@
 /*
- * The Clear BSD License
  * Copyright (c) 2015, Freescale Semiconductor, Inc.
  * Copyright 2016-2017 NXP
  * All rights reserved.
- * 
- * Redistribution and use in source and binary forms, with or without modification,
- * are permitted (subject to the limitations in the disclaimer below) provided
- *  that the following conditions are met:
  *
- * o Redistributions of source code must retain the above copyright notice, this list
- *   of conditions and the following disclaimer.
- *
- * o Redistributions in binary form must reproduce the above copyright notice, this
- *   list of conditions and the following disclaimer in the documentation and/or
- *   other materials provided with the distribution.
- *
- * o Neither the name of the copyright holder nor the names of its
- *   contributors may be used to endorse or promote products derived from this
- *   software without specific prior written permission.
- *
- * NO EXPRESS OR IMPLIED LICENSES TO ANY PARTY'S PATENT RIGHTS ARE GRANTED BY THIS LICENSE.
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ * SPDX-License-Identifier: BSD-3-Clause
  */
 #ifndef _FSL_EWM_H_
 #define _FSL_EWM_H_
@@ -40,7 +14,6 @@
  * @addtogroup ewm
  * @{
  */
-
 
 /*******************************************************************************
  * Definitions
@@ -64,10 +37,10 @@ typedef enum _ewm_lpo_clock_source
 #endif /* FSL_FEATURE_EWM_HAS_CLOCK_SELECT */
 
 /*!
-* @brief Data structure for EWM configuration.
-*
-* This structure is used to configure the EWM.
-*/
+ * @brief Data structure for EWM configuration.
+ *
+ * This structure is used to configure the EWM.
+ */
 typedef struct _ewm_config
 {
     bool enableEwm;           /*!< Enable EWM module */
@@ -135,7 +108,7 @@ extern "C" {
  *
  * @param base EWM peripheral base address
  * @param config The configuration of the EWM
-*/
+ */
 void EWM_Init(EWM_Type *base, const ewm_config_t *config);
 
 /*!
@@ -144,7 +117,7 @@ void EWM_Init(EWM_Type *base, const ewm_config_t *config);
  * This function is used to shut down the EWM.
  *
  * @param base EWM peripheral base address
-*/
+ */
 void EWM_Deinit(EWM_Type *base);
 
 /*!
@@ -187,7 +160,7 @@ void EWM_GetDefaultConfig(ewm_config_t *config);
  */
 static inline void EWM_EnableInterrupts(EWM_Type *base, uint32_t mask)
 {
-    base->CTRL |= mask;
+    base->CTRL |= (uint8_t)mask;
 }
 
 /*!
@@ -202,7 +175,7 @@ static inline void EWM_EnableInterrupts(EWM_Type *base, uint32_t mask)
  */
 static inline void EWM_DisableInterrupts(EWM_Type *base, uint32_t mask)
 {
-    base->CTRL &= ~mask;
+    base->CTRL &= (uint8_t)(~mask);
 }
 
 /*!
@@ -222,7 +195,7 @@ static inline void EWM_DisableInterrupts(EWM_Type *base, uint32_t mask)
  */
 static inline uint32_t EWM_GetStatusFlags(EWM_Type *base)
 {
-    return (base->CTRL & EWM_CTRL_EWMEN_MASK);
+    return ((uint32_t)base->CTRL & EWM_CTRL_EWMEN_MASK);
 }
 
 /*!
@@ -231,7 +204,7 @@ static inline uint32_t EWM_GetStatusFlags(EWM_Type *base)
  * This function resets the EWM counter to zero.
  *
  * @param base EWM peripheral base address
-*/
+ */
 void EWM_Refresh(EWM_Type *base);
 
 /*@}*/
